Title: 부분 문자열 중 팰린드롬의 수, Time: 305ms, Memory: 12MB, Status: Passed - CodeTree

+++ 231118/부분 문자열 중 팰린드롬의 수/counting-palindrome-substrings.java
@@ -44,7 +44,7 @@ public class Main {
 			}
 		}
 		
-		int ans = 0;
+		long ans = 0;
 		for(int i = 0; i < n ;i++) {
 			int max = 2 * A[i] + 1;
 			int l = max / 2;
